Move some files around to prepare for a general solution for donation events.
